requirements.txt
asgiref==3.7.2
Django==4.2.5
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
drf-yasg==1.21.7
factory-boy==3.3.0
Faker==19.6.2
inflection==0.5.1
iniconfig==2.0.0
packaging==23.1
Pillow==10.0.1
pluggy==1.3.0
psycopg2-binary==2.9.7
PyJWT==2.8.0
pytest==7.4.2
pytest-django==4.5.2
pytest-factoryboy==2.5.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
PyYAML==6.0.1
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.8.0
uritemplate==4.1.1
